实现：定时拉区块

+++ src/main/java/com/webank/fabric/node/manager/accessory/scheduler/ScheduleProperties.java
@@ -18,8 +18,8 @@ public class ScheduleProperties {
     private BigInteger pullBlockInitCnts;
     private Long pullBlockTaskFixedDelay;
     private boolean isDeleteInfo;
-    private Long transRetainMax;
-    private Long blockRetainMax;
+    private BigInteger transRetainMax;
+    private BigInteger blockRetainMax;
     private String deleteInfoCron;
     private String statisticsTransDailyCron;
 }
